Add simple cli commands
